requirements.txt
astroid==2.2.5
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
docutils==0.14
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
packaging==19.0
pip-review==1.0
pkginfo==1.5.0.1
psutil==5.6.1
psycopg2==2.8
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
python-dotenv==0.10.1
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
SQLAlchemy==1.3.2
tqdm==4.31.1
twine==1.13.0
typed-ast==1.3.1
urllib3==1.24.1
webencodings==0.5.1
wrapt==1.11.1
